+++ src/main/java/um/tesoreria/core/hexagonal/usuario/domain/ports/in/FindUsuarioByLoginUseCase.java
@@ -1,0 +1,8 @@
+package um.tesoreria.core.hexagonal.usuario.domain.ports.in;
+
+import um.tesoreria.core.hexagonal.usuario.domain.model.Usuario;
+import java.util.Optional;
+
+public interface FindUsuarioByLoginUseCase {
+    Optional<Usuario> findUsuarioByLogin(String login);
+}